+++ pkg/service/usage.go
@@ -26,7 +26,7 @@ func (s *UsageService) GetUsage(startDate, endDate float64) {
 	if err != nil {
 		log.Fatalln("err: ", err)
 	} else {
-		jsonPayload, err := json.Marshal(res.GetPayload())
+		jsonPayload, err := json.MarshalIndent(res.GetPayload(), "", "\t")
 		if err != nil {
 			log.Fatalln("err: ", err)
 		}