+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 **BREAKING CHANGES**
 - Fix `--principle-id` flag to be `principal-id`
+- Remove MasterAccountCreds from dce.yml. Use default AWS CLI creds instead
+- Indent JSON output from CLI commands
 
 ## v0.1.0